added docstrings, changed tests

+++ api/src/main/java/cz/muni/fi/pa165/dto/ActorDto.java
@@ -15,7 +15,7 @@ import java.util.Set;
 public class ActorDto {
     Long id;
     String fullName;
-    Double height;
+    int height;
     LocalDate birthDate;
     LocalDate deathDate;
 }
